requirements.txt
python-dotenv
discord-py-interactions
motor
